auth_device/auth/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==6.0.1
django-rest-swagger==2.2.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.14
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
openapi-codec==1.3.2
packaging==26.0
psycopg2==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simplejson==3.20.2
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
